refactor: 패키지 명을 가급적 한 단어를 사용해 수정
