+++ src/main/java/roomescape/common/config/WebConfig.java
@@ -5,7 +5,7 @@ import org.springframework.context.annotation.Configuration;
 import org.springframework.web.method.support.HandlerMethodArgumentResolver;
 import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-import roomescape.presentation.argument_resolver.LoginMemberArgumentResolver;
+import roomescape.presentation.resvoler.LoginMemberArgumentResolver;
 import roomescape.presentation.interceptor.AdminAuthInterceptor;
 
 @Configuration
