+++ src/main/java/roomescape/presentation/resvoler/LoginMemberArgumentResolver.java
@@ -1,4 +1,4 @@
-package roomescape.presentation.argument_resolver;
+package roomescape.presentation.resvoler;
 
 import jakarta.servlet.http.HttpServletRequest;
 import org.springframework.core.MethodParameter;
